+++ test/unit/labels_test.lua
@@ -22,6 +22,11 @@ function g.test_labels_ok()
     )
 
     t.assert_equals(
+        label_utils.validate_labels("field", {labels = box.NULL}),
+        true, "labels of type box.NULL valid"
+    )
+
+    t.assert_equals(
         label_utils.validate_labels("field", {labels = {}}),
         true, "labels of type table valid"
     )
@@ -52,6 +57,12 @@ function g.test_labels_error()
         labels = function()
         end
     })
+    check_error({labels = {["label"] = 1}})
+    check_error({labels = {["label"] = box.NULL}})
+    check_error({labels = {["label"] = function() end}})
+    check_error({labels = {["label"] = true}})
+    check_error({labels = {["label"] = {}}})
+
     check_error({labels = "asdfsd"})
     check_error({labels = {[""] = ""}})
     check_error({labels = {[""] = "1234"}})